QHttpServerRouterRule: fix passing of RouterHandler
The implementation suggested that the intent was to take RouterHandler
by universal reference, however, RouterHandler is not a deduced type
in the QHttpServerRouterRule ctors, so RouterHandler&& is merely an
rvalue reference, and std::forward<RouterHandler> is a no-op. The
current signature therefore only works when temporaries are passed.

This is not desireable, so fix by taking RouterHandler (a
std::function) by value and std::move()ing it into place.

Pick-to: 6.4
Task-number: QTBUG-100868
Change-Id: Ib663bf2794d30270c6c97e5b2fca1c5e864dbd54
Reviewed-by: Mårten Nordheim <marten.nordheim@qt.io>
Reviewed-by: Jesus Fernandez <jsfdez@gmail.com>