Fix memleak when upgrading to websocket
Normally the lifetime of a QTcpSocket is handled by the connections
established in the QAbstractHttpServerPrivate::handleNewConnections()
call.
However, when a QTcpSocket needs to be upgraded to web socket, its
lifetime will be managed by an instance of QWebSocketServer. When an
instance of QWebSocket is created, it wraps the underlying QTcpSocket.
On destruction, such QWebSocket calls QTcpSocket::disconnect(), which
removes all the connections that were previously established in
QAbstractHttpServerPrivate::handleNewConnections(). As a result, the
QHttpServerRequest object is leaked.

In practice, when the TCP socket is upgraded to websocket, we no longer
need to handle its lifetime in the QAbstractHttpServerPrivate, and we
also no longer need the QHttpServerRequest instance. So this patch
explicitly removes all the outdated connections and deletes the request
before upgrading the socket to websocket.

Fixes: QTBUG-104538
Pick-to: 6.4
Change-Id: I518fe50d7eba4207d50eb0eda5b570784f2a4627
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Ievgenii Meshcheriakov <ievgenii.meshcheriakov@qt.io>