Do not use QMetaTypeId2
We want to know whether the type is usable with QVariant - for that we
are only interested in whether we can copy it. So check that directly.

For addConverter, there should not be a need for any precondition.

Change-Id: I66dc7f4a55a25514aa4439c40260b4e80e9cfed8
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Marc Mutz <marc.mutz@qt.io>
(cherry picked from commit 2cb57fa15fbef1787f41c534ebb7388221ae6d7e)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>