QHttpServerResponder: Don't allow to override HTTP version
The version cannot be decided by the handler code alone and
depends on the request version that was used by the client.

The server should be able to choose different versions for
future HTTP/2 support, so having the handler code specifying
version is not helpful.

Hardcode the version to 1.1 for now.

Task-number: QTBUG-100868
Task-number: QTBUG-95069
Pick-to: 6.4
Change-Id: I29f0fe629c3ceee39dac3d55aa3eb6a785774efa
Reviewed-by: Mårten Nordheim <marten.nordheim@qt.io>