Doc: Update overview documentation
Move separate overview page to landing page - this is
the place where people should get a quick overview of what
the module is for, so IMO it better fits here.

While at it, linkify all referenced methods and signals.
Add also details on how to link to the module to this page,
and add a link to the logging page to the reference section.

Pick-to: 6.4
Change-Id: I05ac9a4d519aa96d345703c92382de6795d130c2
Reviewed-by: Konrad Kujawa <konrad.kujawa@qt.io>
Reviewed-by: Alex Blasche <alexander.blasche@qt.io>
Reviewed-by: Edward Welbourne <edward.welbourne@qt.io>