QHttpServerRequest: Handle requests with "Expect: 100-continue" header
RFC 2616 specifies that the server upon receiving a request with
Expect header set to 100-continue must either reply with
100 (Continue) status code or with a final status code.
This patch modifies the behavior of the server to reply with
100 (Continue) status code when the request contains a body.
This makes it possible for clients to use the Expect header
without having to wait for a timeout before sending the request
body.

The change was manually tested using CURL. Unfortunately
QNetworkAccessManager does not handle Expect header correctly
so it cannot be used for and automated test.

Fixes: QTBUG-107854
Change-Id: I75d8ad13181c108c7405c9de82860b2f2272cd3f
Reviewed-by: Mårten Nordheim <marten.nordheim@qt.io>