Change QHttpServerResponder to be suitable for both HTTP 1.1 and 2
The QHttpServerResponder class is too low level and HTTP 1.1 specific
to be used for both HTTP 1.1 and 2. Remove the low level functions and
add member functions for sending data without having to know the
total length of the data before sending it.

To respond with data with length not known before response is
generated, call one of the writeBeginChunked functions, followed by
as many calls to writeChunk as needed, and finish with a call to one
of the writeEndChunked functions.

The HTTP 1.1 implementation has been moved from QHttpServerResponder
to QHttpServerResponderPrivate.

Fixes: QTBUG-124866
Change-Id: Ie4dd6982bbbb14a5403ebcdfe85626deb5fb2cc5
Reviewed-by: Marcus Tillmanns <marcus.tillmanns@qt.io>