Add a benchmark for QHttpServer
It's quite simple, we just download a 100MB payload from ourselves and
discard all the bytes.

The default qbytearray non-TLS case takes around 250ms to complete on my
machine, while the TLS case clocks in at around 800ms.

Some of this benchmark will measure the time it takes to read as well,
but even using an external process to perform the download we would
still have to wait for it to receive all the data.

Pick-to: 6.8
Change-Id: I0fb700b2f3ab5d73a81f673688ec9404a6749daa
Reviewed-by: Mate Barany <mate.barany@qt.io>