Http/1: Call write() less often
We should try to write as much as possible in one go.
If we add an unbuffered mode then each write may result in a direct
system call. But even with buffered mode it depends on if it's a TLS
connection and the state of the internal ring buffer if it'll be split
up or not.

Change-Id: I337a05884e95d100705663a27ecdece167b1ea10
Reviewed-by: Edward Welbourne <edward.welbourne@qt.io>
Reviewed-by: Lena Biliaieva <lena.biliaieva@qt.io>
(cherry picked from commit 181e133d44f3d69d13911220303bdec0f4d5c087)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>