QHttpServerResponse: Fix constness of status in implementation
The status parameter used in QHttpServerResponse constructors
are not const in header file, but const in the implementation.
This compiles on all compilers in CI, but does not work when
generating documentation. Status is not supposed to be const
because it is passed by value. Fix the implementation as well.

Also set warning limit to zero to enforce zero warnings in CI.

Task-number: QTBUG-125859
Pick-to: 6.8
Change-Id: Ia34831bd6210ca027666840056693d89562f97a3
Reviewed-by: Paul Wicking <paul.wicking@qt.io>