Add constant getter for router to QHttpServer
The constant getter returns a pointer to a constant QHttpServerRouter.

This interface was chosen bacause router is a member of QHttpServer and
the returned pointer is thus not just a reference to another object but
a pointer to a member variable. Therefore router should be considered
part of the state of QHttpServer and making it changeable could break
const-correctness (and would require a const_cast).

Change-Id: Ibaed13b6193babe2a0869feaf6a814963afd4eee
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>
(cherry picked from commit c82e83825ae38b969fcc217a0f98497d79d04ac8)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>