Fix memory leaks in test code of tst_qabstracthttpserver
Create QTcpServer objects in stack rather than in the heap to simplify memory ownership.
Normally QTcpServer created in the heap is not leaking, because it becomes a child of QAbstractHttpServer in QAbstractHttpServer.bind(). However parenting does not happen if bind() fails. So QTcpServer objects were leaking in tests where bind() was supposed to fail.
As the simplest fix, it was decided to create those QTcpServer objects in the stack.

Change-Id: I19788dff4c55deacf242be03871512c9c86296de
Reviewed-by: Mate Barany <mate.barany@qt.io>
Reviewed-by: Øystein Heskestad <oystein.heskestad@qt.io>