Shorten local socket path in autotest
The limitation on Unix domain socket paths on Linux is 107 characters.
After c617cc95934ae3c0896082d61a88487b34cf96be the temp directory
handling on Android must have changed to go past that limitation. When
testing this the socket path on my device was

/data/user/0/org.qtproject.example.tst_qhttpservermultithreaded/cache/
tst_qhttpservermultithreaded_local11757

...which is too long and therefore the listening on the socket will
fail.

The simple quick solution to get everything working is to just shorten
the socket name a bit. Combined with the application PID it should still
be unique and not cause any other issues.

Pick-to: 6.10 6.9
Fixes: QTBUG-137849
Change-Id: Ib606ea315d7a929a37a0639c07ed08b474b54812
Reviewed-by: Assam Boudjelthia <assam.boudjelthia@qt.io>
Reviewed-by: Øystein Heskestad <oystein.heskestad@qt.io>