Sequential QIODevice in tst_qhttpserver now emits readChannelFinished
The mock QIODevice SequentialIODevice in test now emits
readChannelFinished() when all data is available.

Now it also adds data multiple times with a delay instead
of having it available from the time of construction.

A new test is added to ensure that large amounts of data
can be read.

Task-number: QTBUG-137330
Pick-to: 6.10 6.9 6.8
Change-Id: Ibea47feefeffebd64a43a1a1ec5f00687b246ea6
Reviewed-by: Matthias Rauter <matthias.rauter@qt.io>