Add tests for missing features in HTTP/1.0 support
Add tests that verify that the response has the same
version as the request.

Verify that chunked transfer encoding is not used for
HTTP/1.0, because it is not supported. QSKIP tests
using sequential device because it sometimes closes
the connection too soon. The fixes for these issues
are in an upcoming patch that will not be backported.

Finally refactor the client used in the test.

Task-number: QTBUG-138410
Pick-to: 6.11 6.10 6.8
Change-Id: Ibc45520e30b02fe7f4e15848114b28ed9bcce03f
Reviewed-by: Matthias Rauter <matthias.rauter@qt.io>