Complete HTTP/1.0 support
Make QtHttpServer support writeChunk() and write() using sequential
QIODevice for HTTP/1.0. Close the socket when all data is sent
because HTTP/1.0 does not have chunked transfer encoding support.

The socket is closed using disconnectFromHost() because close()
closes immediately before transmitting the remaining contents. Also
use invokeMethod() using QueuedConnection to ensure all data is
processed before closing.

For QLocalSocket use disconnectFromServer(), even though it
is the server, because this is the function that closes the
connection after sending all data, unlike close() which closes
abruptly.

Respect the incoming "Connection" header, which is by default
"close" for HTTP/1.0 and "keep-alive" for HTTP/1.1. If the
length is unknown for HTTP/1.0 set it to "close".

Send the "Connection" header with value "keep-alive" for tests
where the connection is reused.

Fixes: QTBUG-138410
Change-Id: Ia99db97557cff83937c0fa98f166b414d2c51ba3
Reviewed-by: Edward Welbourne <edward.welbourne@qt.io>
(cherry picked from commit 4fd19291a3a6c77991b0905785ec628868c12604)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>