Clear connnection cache of QNetworkAccessManager to prevent flaky tests
Clear connection cache between tests to stop a bug in QNAM from causing
reused HTTP/2 connections to sometimes become stuck in tests on Windows
11. This is a temporary fix to flakyness in tst_qhttpserver until
the bug in QNAM is resolved.

Increase the transfer timout in tst_QHttpServer::timeoutConnection
because the full TLS handshake have to be done for all https/1.1 and
http/2 tests now.

Task-number: QTBUG-142127
Change-Id: I9f61f3e38490423436d38dd7db1939e5b5c9852e
Reviewed-by: Dennis Oberst <dennis.oberst@qt.io>