Limit the number of connections for each IP address
Count the number of active connections for each IP address and refuse
incoming connections when that amount is reached. The number of
connections can be configured in QHttpServerConfiguration. If
set to 0, the limitation is disabled. By default this is disabled.

For HTTP/1.* HTTP status 429 is returned.
For SSL connections the connection is closed during SSL handshake

[ChangeLog] Added configurable limit to the number of active connections
for each IP address.

Task-number: QTBUG-142563
Change-Id: Ie1be2539738a9eff059d21f8fb190fb8c6f0ca60
Reviewed-by: Nils Petter Skålerud <nils.petter.skalerud@qt.io>
Reviewed-by: Mårten Nordheim <marten.nordheim@qt.io>